REFACTOR(SinglePlanScreen): 재스케줄링 버튼 위치이동
