+++ presentation/src/main/res/drawable/icon_reschedule.xml
@@ -6,28 +6,28 @@
   <path
       android:pathData="M10,21H5a2,2 0,0 1,-2 -2v-9h18m-6,-6V2m0,2v2m0,-2h-4.5M3,10V6a2,2 0,0 1,2 -2h2m0,-2v4m14,4V6a2,2 0,0 0,-2 -2h-0.5"
       android:strokeLineJoin="round"
-      android:strokeWidth="1.5"
+      android:strokeWidth="1"
       android:fillColor="#00000000"
       android:strokeColor="#000"
       android:strokeLineCap="round"/>
   <path
       android:pathData="M21.667,16.667C21.048,15.097 19.635,14 17.99,14c-1.759,0 -3.253,1.255 -3.794,3"
       android:strokeLineJoin="round"
-      android:strokeWidth="1.5"
+      android:strokeWidth="1"
       android:fillColor="#00000000"
       android:strokeColor="#000"
       android:strokeLineCap="round"/>
   <path
       android:pathData="M19.995,16.772H21.4a0.6,0.6 0,0 0,0.6 -0.6V14.55m-7.666,4.783C14.953,20.903 16.366,22 18.01,22c1.759,0 3.253,-1.255 3.794,-3"
       android:strokeLineJoin="round"
-      android:strokeWidth="1.5"
+      android:strokeWidth="1"
       android:fillColor="#00000000"
       android:strokeColor="#000"
       android:strokeLineCap="round"/>
   <path
       android:pathData="M16.005,19.228H14.6a0.6,0.6 0,0 0,-0.6 0.6v1.622"
       android:strokeLineJoin="round"
-      android:strokeWidth="1.5"
+      android:strokeWidth="1"
       android:fillColor="#00000000"
       android:strokeColor="#000"
       android:strokeLineCap="round"/>
